requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
config==0.5.1
constantly==15.1.0
cryptography==41.0.4
cssselect==1.2.0
discord.py==2.3.2
django-commands==0.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
filelock==3.12.4
Flask==3.0.0
frozenlist==1.4.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.2
parsel==1.8.1
pluggy==1.3.0
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
PySocks==1.7.1
pytest==7.4.2
pytest-asyncio==0.21.1
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
Scrapy==2.11.0
selenium==4.14.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tldextract==5.0.0
tomli==2.0.1
trio==0.22.2
trio-websocket==0.11.1
tweepy==4.14.0
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.8.0
urllib3==2.0.7
validators==0.22.0
w3lib==2.1.2
Werkzeug==3.0.0
wsproto==1.2.0
yarl==1.9.2
zope.interface==6.1
